Add Report interface to hide pterm and allow multiple output modes
Currently just adds a layer between pterm and the calling code but will
eventually allow machine-readable output or non-colour versions (maybe
for CI mode?). Side effect of this is that it makes the rest of the code
a bit cleaner as the pterm API has a lot of options available and is
therefore slightly complex.

In addition to this, the reporter interface is now properly mocked in
the unit tests and doesn't require pterm.DisableOutput(), which is much
nicer. In future we could even add more detailed expectations on this
mocked report to verify that behaviour.