Abstract away client instantiation and use Reporter for error logging
The SSH connection errors were previously just doing a fmt.Println with the error message which was ugly compared to the rest of the pretty printing.
This commit changes that so the Reporter.Check is used to live-log the progress of setting up SSH and checking the Vault API can be accessed.