Get PKI Monitor working
Had to be more careful about the naming of the policies, then ensure the
intermediate certificate request process was correct as there were a few
mistakes there. The proper TPP and Cloud connectors are used instead of
the fake one, and I had to pass a zone in so the API knew which CA
template to use. I also specified the certificate Subject info in the
role as it is needed there. Finally I updated the policies at the end
with the role, so that the import and enforcement all worked properly.