Add idempotency to apply command
Checks if plugin already in catalog and accordingly either reenables and
reloads, skips the enable step, or enables as usual.

Similarly with the mount plugin step, we check if the plugin is already
mounted and skip the step if so.

I also moved these steps and the download step out of the install plugin
step so it has less responsilbity and added tests.