Add printf style methods to Reporter to avoid fmt.Sprintf everywhere
Previously the codebase was littered with check.Error(fmt.Sprintf())
which is a bit redundant so I added a check.Errorf function to do this
transparently.

This added some problems with mocks for a while because I couldn't get
the variadic arguments to match the mock expectation without listing
each argument unvariadically. However I found a solution in the mock
autogeneration tool mockery, which is to set --unroll-variadic=false.
The effect of this is that the arguments get passed into the mock call
matcher as a slice instead of individual arguments, so mock.Anything can
match it with only one.