Zone required for pki-backend, plugins both request test certs as check
Zone is required for pki-backend but optional for pki-monitor so I
amended as such.

Instead of doing fairly redundant checks in the plugin Verify functions,
they both now call a common RequestCertificate function to exercise the
new backend.

I've also renamed the vagrant test env to make it clear that it's an HA
cluster using integrated raft storage. This should make it easier to
have multiple vagrant envs for testing. I updated the vvw.hcl file to
install both plugins.