Add Kubernetes Helm test environment and support pre-installed plugin binaries
For the Kubernetes test environment I have used a Docker image that has the plugin binaries pre-downloaded and baked into the image.
In this case the VVW tool does not need to have SSH access to install the plugins, and it can just skip this step.
I have allowed the vvw.hcl file to not specify any SSH blocks and added an Info message to reflect this.
I also added a custom filename parameter in case the naming convention is not followed.