Implement generate-config steps for plugins
Allows PluginImpls to implement their own prompt and HCL writing logic.
Also required the plugin lookup logic to change a bit so it can be reused between apply and generate-config commands.
In particular there is now a dedicated ParseConfig method that gets called from the config.NewConfig function, instead of it being part of the LookupPlugin call.
This is because the generate-config command also needs to "look up" the PluginImpl, but does not need to parse any config.

Various plugin-related structs have got WriteHCL methods for serialising themselves into HCL, and there are some GenerateConfigAndWriteHCL methods knocking about which run through the complicated question prompts and produce the relevant config structs before calling their WriteHCL methods.
To aid in the complicated question prompt descriptions, I have added a questions package which allows a more declarative way of specifying the questions to be asked, and the branching based on previous answers.