Add more of pki-monitor config generation logic and use AnswerQueue instead of string slice
The string slice indexing was getting a little messy with all the branching in pki-monitor so I added an AnswerQueue package to allow pushing/popping from the answers without having to manage the current index or do really deeply nested if statements.
Now, instead of accessing the answers based on their index (which changes depending on the branch taken, due to the possible different length branch executions), the answers just need to be "Pop"ped out in the order that the questions were asked.
In other words, as long as the answer parsing makes the same number of calls to Pop as there were questions asked, things should work out fine.

Having made this simplification, the certs and policies parts of the pki-monitor config were implemented.

I also ran generate-mocks due to the recent changes to the PluginImpl interface.