Refactor questions package for testability and remove AnswerQueue jank
The previous version of the questions package was working fine but had a couple of issues.

Firstly the hardcoded dependency on promptui meant that it was hard to test.
This was because it would always ask for stdin and there was no easy way to mock out this dependency.

Secondly, the AnswerQueue mechanism and the need to "pop" out answers in the same order the questions were asked was fragile at best.
This only exacerbated the need for testing, as it was easy to introduce logic errors.
The only way to test this previously was to manually run the app and try the wizard, hoping to avoid a panic.
Hopefully this new structure will allow a mock implementation of the Questioner interface, which will mean the expected question list and branch paths can be verified automatically.

In addition to being more testable, the structure of the question slices has changed to now revolve around named questions (via a map).
This means that the answers are extracted from each question individually, by referencing its map key.
This avoids having to "pop" the answers out in the right order, and now means they can be referenced like variables.