Add new optional_config block to pki-backend plugin
It was implemented and working for the pki-monitor plugin so I thought I'd add it to the other venafi plugin too.
There was also a bug in the GetAsMap() function where it would panic if called when optionalConfig==nil.
I added a nil check to catch this and return and empty map.
Lastly I had to update the generate_config tests to include the new question about optional parameters.
These are already tested in the container-pki-monitor_optional tests so I said "No" to the optional config in all of the other tests.