Rename generic `Plugin` struct to `PluginConfig`, and `PluginImpl` interface to `Plugin`
It was a bit confusing having "Impl" in an interface when it is actually the opposite.
Therefore I renamed `Plugin` to `PluginConfig`, which is similar to the existing `VaultConfig` struct.
This allowed `PluginImpl` to become `Plugin` which feels more apt.
Hopefully the name `Plugin` and its "interface" kind will sufficiently represent that it's an abstract interface over a plugin.