Refactor downloader into granular functions and call from plugin implementation
While trying to add the Redis Enterprise Vault plugin to the tool, it was discovered that the Git releases use a different format.
The previously assumed packaging format of a series of ZIP files containing the plugin its SHA, is not used.
Therefore it is necessary to specialise the plugin downloading process depending on the plugin.
To do this, I changed the GetDownloadURL method on the Plugin interface to be DownloadPlugin, and moved the download logic inside it.
This means each plugin can implement its own logic for downloading, in addition to determining the download URL.
However, some of the steps are shared between the plugins, in particular downloading a file over HTTP, and extracting a SHA from the plugin.
I therefore split the previously simple DownloadPluginAndUnzip method into more fine-grained methods that the Redis Enterprise plugin implementation can pick and choose from.
I left the do-it-all DownloadPluginAndUnzip as a helper method to handle all of the steps as I think it should be the default for plugin implementations if they can.

While making these changes I also noticed that nothing uses the PluginDownloader mock anymore so I refactored the package to remove the interface/private struct pattern and just use functions.
I don't think this needs fancy dependency injection anymore so it makes sense to leave it out.