Add secret() config function to retrieve Vault secrets in config file
Adds to the env() function which allows config values to be specified in terms of environment variables.
The secret() function now accepts a Vault path and field key and retrieves a string value from Vault.
This should work for any value except the API address and Vault token, as these are parsed separately prior, and are obviously prerequisite for looking up any other value.
However these two values will still work with the env() function as before.

The config parsing process has had to be modified a bit to allow a two stage parse, and to allow passing a Vault API client back to the config parser midway through the config parsing process.

This commit has an element of WIP as the error handling in the Apply function needs fixing up, and also it might be prudent to add some unit tests to the config parsing to exercise the new secret() function.