Fix logging in build arch check
A check always needs to end in either a success or error, otherwise the spinner will interfere with all subsequent logging statements.
In this case, one was missing.
To fix it, I added a check.Success call at the end to ensure the check was terminated properly.
I also tweaked some wording while I was there.